requirements.txt
django>=4.2
psycopg2-binary>=2.9
django-import-export>=4.0
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.1
